Rename: noticeService 메소드명 수정
noticeService 메소드명 수정했습니다.
