+++ src/main/java/com/mycompany/metamong/controller/NoticeController.java
@@ -8,7 +8,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import com.mycompany.metamong.dto.NoticeDto;
+import com.mycompany.metamong.dto.notice.NoticeDto;
 import com.mycompany.metamong.service.NoticeService;
 
 import lombok.extern.slf4j.Slf4j;
@@ -23,7 +23,7 @@ public class NoticeController {
 	
 	@GetMapping("/noticeList")
 	public String noticeList(Model model) {
-		List<NoticeDto> list = noticeService.selectNoticeList();
+		List<NoticeDto> list = noticeService.getNoticeList();
 		model.addAttribute("list", list);
 		
 		return "notice/noticeList";
