+++ src/main/java/com/mycompany/metamong/daoMain/NoticeDao.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
 
-import com.mycompany.metamong.dto.NoticeDto;
+import com.mycompany.metamong.dto.notice.NoticeDto;
 
 @Mapper
 public interface NoticeDao {
