+++ src/main/java/com/mycompany/metamong/dto/notice/NoticeDto.java
@@ -1,4 +1,4 @@
-package com.mycompany.metamong.dto;
+package com.mycompany.metamong.dto.notice;
 
 import java.util.Date;
