+++ src/main/java/com/mycompany/metamong/service/NoticeService.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.mycompany.metamong.daoMain.NoticeDao;
-import com.mycompany.metamong.dto.NoticeDto;
+import com.mycompany.metamong.dto.notice.NoticeDto;
 
 @Service
 public class NoticeService {
@@ -14,7 +14,7 @@ public class NoticeService {
 	private NoticeDao noticeDao;
 	
 
-	public List<NoticeDto> selectNoticeList() {
+	public List<NoticeDto> getNoticeList() {
 		return noticeDao.selectNoticeList();
 		
 	}
